Revert "This adds rookout to tron (TRON-1764) (#875)"
This reverts commit f00e56ce0a7fc90bdd5d5497c148e44bfae16beb.
