+++ requirements-dev.txt
@@ -5,6 +5,7 @@ cfgv==2.0.1
 entrypoints==0.3
 flake8==3.7.9
 identify==1.4.9
+importlib-resources==1.0.2
 iniconfig==1.1.1
 isort==4.3.18
 lazy-object-proxy==1.3.1
