+++ requirements.txt
@@ -20,11 +20,9 @@ cryptography==2.7
 dataclasses==0.6
 DateTime==4.3
 decorator==4.4.0
-distro==1.7.0
 docker==4.1.0
 docutils==0.15.2
 ecdsa==0.13.3
-funcsigs==1.0.2
 future==0.18.1
 google-auth==1.23.0
 http-parser==0.9.0
@@ -32,7 +30,6 @@ humanize==0.5.1
 hyperlink==19.0.0
 idna==2.8
 importlib-metadata==1.3.0
-importlib-resources==5.4.0
 incremental==17.5.0
 ipdb==0.13.2
 ipython==7.8.0
@@ -49,14 +46,14 @@ kubernetes==17.17.0
 lockfile==0.12.2
 MarkupSafe==1.1.1
 mock==3.0.5
+more-itertools==8.0.2
 moto==1.3.13
 oauthlib==3.1.0
 parso==0.5.1
 pexpect==4.7.0
 pickleshare==0.7.5
 prompt-toolkit==2.0.10
-protobuf==3.19.4
-psutil==5.9.1
+psutil==5.6.6
 ptyprocess==0.6.0
 py-bcrypt==0.4
 pyasn1==0.4.7
@@ -77,7 +74,6 @@ PyYAML==5.4.1
 requests==2.22.0
 requests-oauthlib==1.2.0
 responses==0.10.6
-rook==0.1.160
 rsa==4.0
 s3transfer==0.2.1
 setuptools==41.4.0
@@ -93,5 +89,5 @@ websocket-client==0.56.0
 Werkzeug==0.16.0
 wrapt==1.11.2
 xmltodict==0.12.0
-zipp==3.1.0
+zipp==0.6.0
 zope.interface==5.1.0
