+++ tron/trondaemon.py
@@ -85,10 +85,7 @@ class TronDaemon:
                 signal.SIGUSR1: self._handle_debug,
             }
             signal.pthread_sigmask(signal.SIG_BLOCK, signal_map.keys())
-            try:
-                self._run_rookout()
-            except Exception:
-                logging.exception("Unable to setup Rookout.")
+
             self._run_mcp()
             self._run_www_api()
             self._run_manhole()
@@ -99,13 +96,6 @@ class TronDaemon:
                 if signum in signal_map:
                     logging.info(f"Got signal {str(signum)}")
                     signal_map[signum](signum, None)
-
-    def _run_rookout(self):
-        # Local import required because logging is not initialized until later
-        from tron.utils import rookout
-
-        log.info("Preparing Rookout SDK")
-        rookout.enable_rookout()
 
     def _run_manhole(self):
         # This condition is made with the assumption that no existing daemon
