+++ yelp_package/bionic/Dockerfile
@@ -22,7 +22,6 @@ RUN apt-get -q update && \
         python3-pip \
         python-tox \
         wget \
-        g++ \
     && apt-get -q clean
 
 ARG PIP_INDEX_URL
